Make the demo more pretty
