+++ tests/data/methods.py
@@ -26,3 +26,14 @@ def sleep_queue_one(seconds=1):
 @task(queue='test_channel')
 def print_hello():
     print('hello world!!')
+
+
+@task(queue='test_channel', bind=True)
+def hello_world_binder(binder):
+    print(f'Values in binder {vars(binder)}')
+    print(f'Hello world, from worker {binder.worker_id} running task {binder.uuid}')
+
+
+@task(queue='test_channel', timeout=1)
+def task_has_timeout():
+    time.sleep(5)
