--- tools/test_methods.py
@@ -1,34 +1,0 @@
-import time
-
-from dispatcher.publish import task
-
-
-@task(queue='test_channel')
-def sleep_function(seconds=1):
-    time.sleep(seconds)
-
-
-@task(queue='test_channel', on_duplicate='discard')
-def sleep_discard(seconds=1):
-    time.sleep(seconds)
-
-
-@task(queue='test_channel', on_duplicate='serial')
-def sleep_serial(seconds=1):
-    time.sleep(seconds)
-
-
-@task(queue='test_channel', bind=True)
-def hello_world_binder(binder):
-    print(f'Values in binder {vars(binder)}')
-    print(f'Hello world, from worker {binder.worker_id} running task {binder.uuid}')
-
-
-@task(queue='test_channel')
-def print_hello():
-    print('hello world!!')
-
-
-@task(queue='test_channel', timeout=1)
-def task_has_timeout():
-    time.sleep(5)
